refactor(session): lift DocxSession bridge into shared Internal core
Moves the integer-handle pool, per-op session-lookup + serialize facade, and
the StringBuilder JSON helpers that lived inside wasm/DocxodusWasm/DocxSessionBridge.cs
into Docxodus/Internal/ as SessionRegistry, DocxSessionOps, DocxSessionJson.
The WASM bridge becomes a thin [JSExport]-attributed shell over DocxSessionOps.

This unblocks a second transport (stdio NDJSON for python-docxodus) sharing
the exact same JSON wire shapes by construction — adding a new DocxSession
op is now one edit in DocxSessionOps + one passthrough per transport,
rather than reimplementing serialization per language client.

Pure refactor — wire format unchanged, all 1411 existing tests pass.
